🎨 style(about/index.tsx): update alt text for download resume image The alt text for the download resume image has been updated to "Download resume" to provide a more descriptive alternative text for accessibility purposes.
🔧 chore(button/styles.tsx): remove transition property
The transition property has been removed from the button styles as it was not being used.

🔧 chore(contact/styles.tsx): update background color, border color, and box shadow
The background color has been updated to var(--gray-900), the border color has been updated to var(--gray-800), and a box shadow with var(--shadow-10) has been added to the contact input styles.

✨ feat(header/data.tsx): add header navigation links
A new file, header/data.tsx, has been added to the project. It exports an array of navigation links for the header component. The links include "About", "Skills", "Experience", "Projects

🔀 refactor(header/index.tsx): remove unused code and dependencies, refactor navLinks
🔀 refactor(header/index.tsx): remove unused code and dependencies, refactor navLinks

- Removed unused code and dependencies such as `Button`, `NavLink`, and `ThemeToggle`.
- Refactored `navLinks` to remove hard-coded values and replaced them with dynamic values.
- Improved code readability and maintainability by removing unnecessary code and dependencies.

🔨 refactor(header/index.tsx): extract NavLinks, MenuButton, and MobileMenuNav components
The code has been refactored to improve readability and maintainability. The NavLinks, MenuButton, and MobileMenuNav components have been extracted into separate components, making the code more modular and easier to understand. This also allows for better reusability of these components throughout the application.

✨ feat(menu-button): add MenuButton component
✨ feat(menu-button/styles): add styles for MenuButton component
✨ feat(mobile-menu-nav): add MobileMenuNav component
The MenuButton component is added, which renders a menu button element. It takes in props for indicating whether the menu is open or closed and a callback function to toggle the menu. The component also includes styles for the button. The MobileMenuNav component is added, which renders the mobile menu navigation. It takes in props for indicating whether the menu is open or closed and a callback function to toggle the menu. The component includes the NavLinks component for rendering the navigation links.

📝 chore(styles): add mobile menu navigation styles and create NavLink component
✨ feat(styles): add MobileMenuNavStyles component
A new file `styles.tsx` has been added to the `src/components/header/mobile-menu-nav` directory. This file contains the styles for the mobile menu navigation component. The `MobileMenuNavStyles` component is defined, which sets the styles for the mobile menu navigation.

✨ feat(nav-link): create NavLink component
A new file `index.tsx` has been added to the `src/components/header/nav-link` directory. This file contains the implementation of the `NavLink` component. The `NavLink` component is a reusable component that renders a navigation link. It accepts props such as `href`, `children`, and `onClick`, and returns a JSX element representing the rendered NavLink component.

🎨 style(nav-link): add styles for navigation links
🎨 style(nav-links): add styles for navigation links component

The `nav-link` component now has styles defined using styled-components. The styles include alignment, positioning, and hover/focus effects. The `nav-links` component now has styles defined as well, which include layout and spacing for the navigation links and additional elements such as a button and a theme toggle.

✨ feat(nav-links): add styles for navigation links
This commit adds a new file `styles.tsx` under `src/components/header/nav-links` directory. The file contains styled components for the navigation links. The `NavLinksStyles` component is a styled `nav` element with flexbox properties to display the links in a row, justify them to the end, and align them in the center. It also sets padding and gap between the links. Additionally, there is a media query that changes the flex direction to column and toggles the display based on the `open` prop when the screen width is less than or equal to 1200px.

🔧 fix(header/styles.tsx): remove unused styles and update link hover effect
🔧 fix(hero/index.tsx): add alt text to scroll down icon
In the header/styles.tsx file, several styles related to the navigation links and mobile menu have been removed as they were not being used. Additionally, the hover effect on the links has been updated to change the color and width of a pseudo-element before the link.

In the hero/index.tsx file, an alt attribute has been added to the scroll down icon image to improve accessibility.

🐛 fix(hero): update padding in responsive styles
The padding in the responsive styles for the hero component has been updated. The padding value has been changed from var(--spacing-l-700) to var(--spacing-m-600) to improve the visual appearance and responsiveness of the component on screens with a maximum width of 1024px.
