+++ src/components/header/data.tsx
@@ -1,0 +1,7 @@
+export const navLinks = [
+	{ href: '#About', label: 'About' },
+	{ href: '#Skills', label: 'Skills' },
+	{ href: '#Experience', label: 'Experience' },
+	{ href: '#Projects', label: 'Projects' },
+	{ href: '#Contact', label: 'Contact' },
+]
