+++ src/components/header/styles.tsx
@@ -26,84 +26,3 @@ export const NavContainer = styled.section`
 	height: 100%;
 	position: relative;
 `
-
-export const NavLinks = styled.nav`
-	display: flex;
-	flex-direction: row;
-	justify-content: flex-end;
-	align-items: center;
-	padding: 0;
-	gap: var(--spacing-m-400);
-
-	@media only screen and (max-width: 1200px) {
-		display: none;
-	}
-`
-
-export const LinkStyles = styled.a`
-	display: flex;
-	align-items: center;
-	text-align: right;
-	position: relative;
-	transition: var(--transition-03);
-
-	&:hover,
-	&:active,
-	&:focus {
-		color: var(--primary-color);
-	}
-
-	&::before {
-		content: '';
-		display: block;
-		position: absolute;
-		bottom: 0;
-		left: 0;
-		width: 0;
-		height: 2px;
-		border-radius: var(--border-radius-sm);
-		background-color: var(--primary-color);
-		transition: var(--transition-03);
-	}
-
-	&:hover::before,
-	&:focus::before,
-	&:active::before {
-		width: 100%;
-	}
-`
-
-export const MenuButton = styled.button`
-	background-color: transparent;
-	display: none;
-
-	@media only screen and (max-width: 1200px) {
-		display: flex;
-	}
-
-	img {
-		height: var(--spacing-m-400);
-		width: var(--spacing-m-400);
-	}
-`
-
-export const MobileMenuNav = styled.nav<{ open: boolean }>`
-	display: flex;
-	flex-direction: column;
-	justify-content: flex-start;
-	align-items: center;
-	gap: var(--spacing-m-400);
-	padding: var(--spacing-m-500);
-	position: fixed;
-	top: 6.7rem;
-	right: 0;
-	height: 100vh;
-	background-color: var(--gray-900);
-	box-shadow: var(--shadow-20);
-	transform: ${(props) => (props.open ? 'translateX(0)' : 'translateX(100%)')};
-	transition: var(--transition-03);
-
-	@media only screen and (min-width: 1200px) {
-		display: none;
-	}
-`
